requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
gitpython==3.1.43
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
langchain-groq==0.1.6
sentence-transformers==3.0.0
faiss-cpu==1.8.0
pydantic==2.7.4
tiktoken==0.7.0
streamlit==1.35.0
requests==2.32.3
groq==0.9.0
